api/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.21.6
Pillow==9.1.0
requests==2.27.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
